backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
sqlalchemy
#psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
ollama
pandas
openpyxl
argon2-cffi
python-docx
pypdf
matplotlib
requests
# Add torch for GPU-accelerated training
--extra-index-url https://download.pytorch.org/whl/cu126
torch